+++ ootrando_overworldmap_hamsda/changelog.md
@@ -5,6 +5,7 @@
 - fixed Sheik at Colossus not showing in ER
 - adjusted position of many locations
 - added regular item captures to ER
+- added default captures for non shuffled locations to ER
 
 ## 3.2.1.0